requirements.txt
# Living Runbooks - Core Dependencies
# Install with: pip install -r requirements.txt

# ============================================
# Core Framework
# ============================================
flask==2.3.3
flask-cors==4.0.0
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
pydantic==2.5.0

# ============================================
# Data Processing
# ============================================
pyyaml==6.0.1
requests==2.31.0

# ============================================
# Logging & Monitoring
# ============================================
structlog==23.2.0

# ============================================
# Optional: LLM Integration
# ============================================
# anthropic==0.18.0
# openai==1.12.0

# ============================================
# Optional: Semantic Search
# ============================================
# sentence-transformers==2.3.1
# torch==2.2.0

# ============================================
# Optional: Git Integration
# ============================================
# gitpython==3.1.42

# ============================================
# Optional: Database (Future)
# ============================================
# sqlalchemy==2.0.25
# psycopg2-binary==2.9.9
# alembic==1.13.1

# ============================================
# Optional: Development & Testing
# ============================================
# pytest==8.0.0
# pytest-cov==4.1.0
# pytest-asyncio==0.23.3
# black==24.1.0
# mypy==1.8.0

# ============================================
# Optional: Redis (for WebSocket caching)
# ============================================
# redis==5.0.1
